pyproject.toml
[tool.poetry]
name = "electonics retail pad"
version = "0.1.0"
description = ""
authors = ["UltimaKIND <minion9@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
pillow = "^11.0.0"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
